added data_assistant and expectation suite logic to abstract
