+++ great_expectations/data_context/data_context/cloud_data_context.py
@@ -1,5 +1,5 @@
 import logging
-from typing import List, Mapping, Optional, Union
+from typing import List, Mapping, Optional, Union, cast
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core import ExpectationSuite
@@ -126,6 +126,92 @@ class CloudDataContext(AbstractDataContext):
             )
         )
 
+    def create_expectation_suite(
+        self,
+        expectation_suite_name: str,
+        overwrite_existing: bool = False,
+        ge_cloud_id: Optional[str] = None,
+        **kwargs,
+    ) -> ExpectationSuite:
+        """Build a new expectation suite and save it into the data_context expectation store.
+
+        Args:
+            expectation_suite_name: The name of the expectation_suite to create
+            overwrite_existing (boolean): Whether to overwrite expectation suite if expectation suite with given name
+                already exists.
+
+        Returns:
+            A new (empty) expectation suite.
+        """
+        if not isinstance(overwrite_existing, bool):
+            raise ValueError("Parameter overwrite_existing must be of type BOOL")
+
+        expectation_suite: ExpectationSuite = ExpectationSuite(
+            expectation_suite_name=expectation_suite_name, data_context=self
+        )
+        key: GeCloudIdentifier = GeCloudIdentifier(
+            resource_type="expectation_suite", ge_cloud_id=ge_cloud_id
+        )
+        if self.expectations_store.has_key(key) and not overwrite_existing:
+            raise ge_exceptions.DataContextError(
+                "expectation_suite with GE Cloud ID {} already exists. If you would like to overwrite this "
+                "expectation_suite, set overwrite_existing=True.".format(ge_cloud_id)
+            )
+        self.expectations_store.set(key, expectation_suite, **kwargs)
+        return expectation_suite
+
+    def delete_expectation_suite(
+        self,
+        expectation_suite_name: Optional[str] = None,
+        ge_cloud_id: Optional[str] = None,
+    ):
+        """Delete specified expectation suite from data_context expectation store.
+
+        Args:
+            expectation_suite_name: The name of the expectation_suite to create
+
+        Returns:
+            True for Success and False for Failure.
+        """
+        key: GeCloudIdentifier = GeCloudIdentifier(
+            resource_type="expectation_suite", ge_cloud_id=ge_cloud_id
+        )
+        if not self.expectations_store.has_key(key):
+            raise ge_exceptions.DataContextError(
+                "expectation_suite with name {} does not exist."
+            )
+        else:
+            self.expectations_store.remove_key(key)
+            return True
+
+    def get_expectation_suite(
+        self,
+        expectation_suite_name: Optional[str] = None,
+        ge_cloud_id: Optional[str] = None,
+    ) -> ExpectationSuite:
+        """Get an Expectation Suite by name or GE Cloud ID
+        Args:
+            expectation_suite_name (str): the name for the Expectation Suite
+            ge_cloud_id (str): the GE Cloud ID for the Expectation Suite
+
+        Returns:
+            expectation_suite
+        """
+        key: GeCloudIdentifier = GeCloudIdentifier(
+            resource_type="expectation_suite", ge_cloud_id=ge_cloud_id
+        )
+        if self.expectations_store.has_key(key):
+            expectations_schema_dict: dict = cast(
+                dict, self.expectations_store.get(key)
+            )
+            # create the ExpectationSuite from constructor
+            return ExpectationSuite(**expectations_schema_dict, data_context=self)
+
+        else:
+            raise ge_exceptions.DataContextError(
+                f"expectation_suite {expectation_suite_name} not found"
+            )
+
     def save_expectation_suite(
         self,
         expectation_suite: ExpectationSuite,
@@ -140,6 +226,8 @@ class CloudDataContext(AbstractDataContext):
             expectation_suite: the suite to save
             expectation_suite_name: the name of this expectation suite. If no name is provided the name will \
                 be read from the suite
+            ge_cloud_id: cloud id for saving expectation suite
+            overwrite_existing: should I over-write the Suite if it already exists?
         Returns:
             None
         """
